Login_Form,новый класс DataManager + раскидал по файлам классы
